vfat: Fix mkcksum argument sizes
In case a function argument is known/fixed size array in C, the argument is
still decoyed as pointer instead ( T f(U n[k]) ~= T fn(U *n) ) and therefore
calling sizeof on the function argument will result in the size of the pointer,
not the size of the array.

The VFAT code contains such a bug, this patch fixes it.

Reported-by: Aaron Williams <Aaron.Williams@cavium.com>
Signed-off-by: Marek Vasut <marex@denx.de>
Cc: Tom Rini <tom.rini@gmail.com>
Cc: Aaron Williams <Aaron.Williams@cavium.com>
Tested-by: Michal Simek <michal.simek@xilinx.com>
Reviewed-by: Joe Hershberger <joe.hershberger@ni.com>

	Taken from 2014.01
Signed-off-by: Eli Nidam <elini@marvell.com>

Change-Id: I77ad6589d0b4bbe5b3c8a0db3cf4fdce8c165833
Reviewed-on: http://vgitil04.il.marvell.com:8080/5786
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Eli Nidam <elini@marvell.com>