fix: alp: pon: compilation issue in Z1 version
Most of the macro definition(Z1, A0) in PON module are too old. They are actually the definition in KW2, in Avanta LP, both of Z1 and A0 has similar new definition, thus need to clean them.
And for Avanta LP, only need Z1/A0 macro when both of Z1 and A0 has the register but content is different.

Change-Id: I577deea6661b1d4a1973d192e41753a56022237c
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/6109
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Yehuda Yitschak <yehuday@marvell.com>