fix: mmc: fixed mmc card scan failure & crash
This patch checks if data == NULL when perform the
memcpy, otherwise it leads to crash when copy to
the destination with NULL address.
This is a bug of SDHCI code, it will leads to crash
in case only send command and pass NULL to "data"
input parameter when calling routine sdhci_send_command.

Fixed NULL pointer access when scan mmc card
JIRA SYSTEMSW-441

Change-Id: Ie919a35f1bda2a0562959acac652117180b79fb4
Signed-off-by: Eli Nidam <elini@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/6873
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/27185