doimage: changed has_extension to ext_cnt in main header
previously the amount of extensions was unknown to bootrom
it would process extensions untill the prolog size reached zero.
know we want to pad the prolog to 128 byte so we allow some
padding at the end of the prolog.
So now we indicate the amount of extension inside the prolog

Also added default padding of boot image to 128 byte to allow
using the same image for uart boot via xmodem transmission

Change-Id: I99e2751dd274d313209986091578f20cc757b47e
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/17492
Reviewed-by: Shadi Ammouri <shadi@marvell.com>
Tested-by: Star_Automation <star@marvell.com>