mbus: arlp: added support for Armada-LP decode window
ArLP uses the same MBUS, but register definitions are different:
 1. register offset calculation method is different
 2. window size is a separate register
 3. the bit offset of window base/remap registers is different
 4. ArLP has no remap high register
In order to reuse the MBUS driver as much as possible, move the
register definitions to a dedicate file (per SoC). mbus driver will
include this header file by macro. To adapt to this approach, MBUS
driver is slightly modified.

Change-Id: I6645710fb69603b06273f24ef14c6fe562a5ad5a
Signed-off-by: Joe Zhou <shjzhou@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/22066
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Hanna Hawa <hannah@marvell.com>