ddr-on-pcie: enabled running ddr-on-pcie using pci switch
This patch allows working with DDR on PCIe over a PCI switch
To allow this we:
- modify the BAR fixup in SPL to account for the new bus and device IDs
  caused by the switch topology
- fix the pci driver to maintian some switch configurations during
  u-boot scan
- Skip the DRAM controller during cfg cycle to retain it's configuration

We also shift the entire bus numbering to start at 1 to be compatible
with linux. linux mvebu_pci driver uses a virtual SW bridge that occupies
bus 0 so our PCI controller moves to bus 1. to avoid changing the bus number
in linux (which stops DDR access) we shift the first bus number to 1 in
u-boot.

Change-Id: I349c8a1e1c5c5cf5b1dfae78ccdc3efdb99e7460
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/23402
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>