clearfog: switch and external phy reset deassert
R87 which is a pull up on the clearfog board will be removed in order to force
reset of the switch and external 1512 phy that is connected to the switch.

The forced reset is with a pull down that is already on the board.

With that MPP41 which previously was set as an input doesn't deassert the reset
signal since R87 pullup is now removed; so MPP41 must be driven high to force
reset deassert.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>