mtd: spi-nand: Add Micron SPI NAND support
This patch is to support Micron SPI NAND. Now it only supports
1Gb/2Gb M79A SPI NAND.

This patch is (1/3) of the Micron's official release for M7XA
SPI NAND chips based on Uboot 2015.07. It is part of SPI NAND
Software Solution v2.0, which can be found by the link below.
https://www.micron.com/resource-details/c9a236c7-ed40-4516-a9c5-a445f57c276f

The spi-nand command support is removed from the orignal Micron
patch since u-boot 2017 reconstructed the command directory. It
will be added by a separate patch later.

The u-boot env support for booting from SPI NAND is removed
from the orignal Micron patch as well. The implementation of
u-boot env in u-boot 2017 depends on the driver model. It will
be added by a separate patch if needed.

Change-Id: I383dd7fa22e49a3b828de431c06e98ed86cb68ba
Signed-off-by: Peter Pan <peterpandong@micron.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/33045
Reviewed-by: Igal Liberman <igall@marvell.com>
Reviewed-by: Haim Boot <hayim@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: Wilson Ding <dingwei@marvell.com>
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39172