icu: add initial support for ICU unit
the Interrupt Consolidation Unit (ICU) is a unit located in CP-110
that recieves wire interrupts from CP-110 units and converts them
to interrupts messages towards the AP, in AP-806 there is a dedicated
unit that recieves these interrupts and converts them to wired SPI
interrupts on the GIC.

This driver impelements a static mapping of all CP-110 functional and
error interrupts to GIC SPI IDs and AP System Error Interrupts (SEI)
and Ram Error Interrupts (REI).

This driver is temporarily located in u-boot, hence no DT support or
proper integration. It should be moved soon firmware

Change-Id: I3fe7563cd1b5efc79be5fcd1ed28fcf10d8f9f5a
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/26341
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>