config: disable LLC for armada-7040-rz
The LLC is disabled to enable working with the AHCI unit.
Due to HW bug, the AHCI unit cannot work in coherent mode so it
must be operated in SW coherency mode requiring cache maintainance.
Since in AP-806-Z1 the LLC doesn't respond to cache maintianance
commands but must be maintained manually, we disable it to allow
working with the AHCI unit in SW coherency mode

Change-Id: Icab568fcf35b2fa63de69d677d9497cad6efda76
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/27263
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>