a3700: define SPL jump address to OSLO image offset
This patch adds OSLO image address Kconfig parameter,
and let SPL jump to OLSO at the end.
In SPL+u-boot case, the OLSO is the u-boot image,
in SPL+ATF+u-boot case, the OLSO is the ATF image.
The default OSLO image start address is set to 0x4100000,
which is the ATF image start address after it is loaded.

Change-Id: I627436e3cefb6735349f0289be72f3e5f969cc87
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/28981
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: Wilson Ding <dingwei@marvell.com>