fix: xenon: run soc init according to DT status
for A8k, Xenon IP on the AP requires special SoC MPP configuration,
and if enabled, it disables the functionality of the used MPP by XENON.
this broke SPI support on the AP.

this patch relocates this configuration from generic soc code, to a
weak SoC implemetation, to be called from xenon init driver.
hence, this configuration will be done only if XENON DT entry is enabled.

Change-Id: I9de1f7e4c11a7f9b6521ac703edc77ff9067fae1
Signed-off-by: Omri Itach <omrii@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/29290
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Haim Boot <hayim@marvell.com>