avs: a3700: get lowest VDD from DTS
The same VDD is used when CPU freq is not more than
600MHZ.
The recommended VDD for Armada-3700 Z1 DDR3 DB board is
different with other boards because of HW limitation.
An AVS property "vdd-min" is added to configure the real VDD used
when CPU freq is not more than 600MHZ flexibly.
The VDD config value should be set according to doc, and consider
real HW limitation, such as the DC/DC limitation in Armada-3700
Z1 DDR3 board.

Change-Id: Id9e78012971816618f9e70082d02d0ca6bdf7e6b
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32012
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>