dts: avs: a3700: add vdd_min property for ddr3 and ddr4 board
The vdd-min is used to set the AVS CPU core VDD, which will be
set to AVS VSET1/2/3 and VSET0(when CPU frequency is not more
than 600MHZ).
The Armada-3700 Z1 DB board has different VDD
because of HW limitation.
Normally the recommended vdd-min is 26(1050mV), but
Armada-3700 Z1 DB board has the HW limitation because of DC/DC
converter, the vdd-min for it should not be less than 31(1108mV).

Change-Id: I3315538fac27458bced37beb81415a350b6ad29d
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32013
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: Star_Automation <star@marvell.com>