dts: a80x0: add new device tree for Armada 8K with Serdes-B connected to CP1
In case when Serdes-B module is plugged to the CP1 (connector 41), the
different serdes and MPP configuration needs to be used. Therefore new dts
should be introduced.

Through the Serdes-B two x3310 PHY can be connected
(DB-88x3310-MODULE_SLIM-1575). The serdes 2 and 3 of A8K are routed to the
first PHY (connected through connector 4 of Serdes-B). While serdes 4 and 5
of A8K are routed to the second PHY (connected through connector 5).

The x3310 PHY in DB-88x3310-MODULE_SLIM-1575 module is connected to the
copper interface RJ45. On the copper interface x3310 PHY supports multiple
speeds such as: 10GBASE-T, 5GBASE-T, 2.5GBASE-T, 1000BASE-T, 100BASE-TX
and 10BASE-Te. Therefore A8K serdeses can be configured in various way.
Introduced dts configures two A8K serdeses to SGMII - one SGMII for each
MAC<->x3310PHY connection. It is also worth mentioning, that x3310 PHY
auto-negotiates link speed automatically without any software and MAC
intervention. The only requirement is to set the MAC to the correct mode
(e.g. if going to use 1000BASE-T, A8K serdes can be configured to SGMII).

The serdes 0 has to be set to 'PHY_TYPE_UNCONNECTED', otherwise the
PHY_TYPE_SGMII2 can't be used for serdes 4.

This dts also reconfigures the MPP 42-43 to xSMI, which are routed to xMDIO
and xMDC signals. Both xMDIO and xMDC are used for XSMI communication
between the MAC and x3310 PHY. They are needed for the x3310 PHY firmware
upgrade.

Using MPP from CP0 and serdeses from CP1 may be a bit confusing, but this
is how the signals are wired in the board: both Serdes from CP1 and MPPs
42-43 from CP0 are routed to the connector 41 (CP_1 Serdes Connector).

To force U-Boot to use this configuration, the ftd_config in u-boot should
be change to '3 - DB-ARMADA-80x0 (SPI-CP1-SERDESB)' via the fdt_config
command.

Change-Id: I87d2198604546fed5af41ad66efce61e6f1e629a
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32550
Reviewed-by: Haim Boot <hayim@marvell.com>
Tested-by: Star_Automation <star@marvell.com>