fix: dts: a70x0: a80x0: name gpio nodes differently for CP1 and CP0
If both gpios nodes for CP1 and CP0 are named in the same way, the u-boot
during some operations like 'usb start' or 'gpio status -a' prints warnings
like:
Device 'gpio@440100': seq 1 is in use by 'gpio@440100'
Device 'gpio@440140': seq 2 is in use by 'gpio@440100'
Those warnings are related to the generic u-boot device probe, which is
called during some of the operations like in examples above. During this
process 'uclass' devices are searched. This process catches that some
uclass devices uses the same seq number. Having different names for gpio
prevent u-boot for printing this kind of warning.

Change-Id: If94ae913f99dc58154f67abb65b6fe65cf6a561f
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32928
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Star_Automation <star@marvell.com>