ahci: add support for multiple sata controllers
Currently the ahci driver doesn't allow working with multiple sata
controllers (only ports).
This patch adds this functionality to the ahci driver and
to the scsi command line.
Technically, the driver allocates memory for the max number of ctrl
and devices(currently 2 ctrl, and 8 devices each).
In each call to the driver functions, if the device doesn't belong
to the current ctrl, the driver will load the suitable ctrl
and exec the cmd.

Change-Id: I064cebec5f550047ff40a0e902767c01c7d907b5
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34036
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>