fix: rtc: a80x0: fix rtc drift
The problem was that, in 80x0, the RTC clock was slow since CP0
RTC clock was used while CP1 RTC was connected to the oscillator.

This patch enables CP0 RTC for 70x0 and CP1 RTC for 80x0.

In addition, it adds the 'status' property to the DT node to allow
enabling the connected RTC per device.

Change-Id: I0e1985070ebd036323f8b2a3f91b6575e3f5d639
Signed-off-by: Haim Boot <hayim@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34621
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Tested-by: Star_Automation <star@marvell.com>