dt: armv8: enlarge the ATF reserve memory area
The suspend to RAM feature for Marvell Armada SoC such as Armada-3700
and Armada-8K trigger SoC warm reset during resume, the ATF binary
will be reloaded and relocated.
More ATF area need to be reserved to avoid instability issue.

This patch is just a temporary WA. It will be replaced by the official
bug-fix in future (JIRA: SYSTEMSW-3316).

Change-Id: I53a1931c59a1133b3540a1a7ef762bdcfe492efe
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/35097
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Reviewed-by: Haim Boot <hayim@marvell.com>