fix: a38x: add warning to eFuse 'set' command
eFuse fields JTAGDIS and BOOTSRC are not really burnt with
'efuse set JTAGDIS' and 'efuse set BOOTSRC' commands but only saved
to internal variables. They are burnt only with command
'efuse set SECURE' when Secure Boot mode is enabled.

Hence, a clarifying message was added together with a message with
warning to not reset a board between these commands.

Change-Id: I13d6825b412f86dafc0a2ca50a9432fefbbedf58
Signed-off-by: Vladimir Abram <vladabr@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/35572
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>