fix: nand: Fix nand RW access wrappers
The inline functions nand_read and nand_write wrap calls
to mtd_read and mtd_write APIs.
The MTD APIs are using an additional parameter for returning
the amount of bytes actually handled by the RW operation.
The wrapper uses a pointer to the same input "len" parameter
for both input and output bytes count when calls to MTD API.
However the MTD functions are zeroing the returning bytes
count at the beginning, effectively zeroing the input/output
buffer length as well.

Change-Id: I62cd20733741ef6da59889ba5be5de4c59d4c43d
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Cc: Scott Wood <oss@buserror.net>
Cc: Stefan Roese <sr@denx.de>
Reviewed-on: http://vgitil04.il.marvell.com:8080/36941
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>