Adding the auto boot capablity at u-boot
Added the u-boot command "get_env". The command will try to search
for the uenv.txt from SDcard/mmc/usb. If the uenv.txt is found on
any of the storage media, it loads the env settings from the uenv.txt
and will try to run the uenvcmd programed by the user in the txt
file. By properly writing the uenv.txt, the h/w can be configured to
auto boot to the file system without touching any of the existing env
variables in the u-boot.

Change-Id: If84e60a3e9952b8335dd966789d80d15556c439a