fix: print the "Model" in the beginning of the boot log
Currently, the "Model" is printed in the middle of the boot log,
so it's not distinguished from the rest of the log.
This patch is removing CONFIG_DISPLAY_BOARDINFO_LATE and setting
CONFIG_DISPLAY_BOARDINFO, so the "Model" printing happens earlier
(on top of boot log instead in the middle).

Change-Id: I2c952115daa030333766be813e80efb960f417b1
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38591
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>