fix: bubt: a70x0: fix bubt from USB to SPI
first_blk_device probes the first block device it finds
and keeps probing all devices until the correct block device
looked for is found (In this case USB).
Therefore the workflow of this function is not suitable
for the bubt command since we don't need to probe all the block
devices now avaiable.
blk_get_device only probes the required device.

Change-Id: I70e063fa7befe289019484c9b8e4f2fa74f9424b
Signed-off-by: Christine Gharzuzi <chrisg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38845
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>