fix: usb: skip access NULL device via pointer check
When USB init, for USB storage a LUN device will be
created, which is taken as a child with status of
inactive of USB storage device.
When run command like "usb dev 0", the LUN device is
set active, which is not a real USB device, and the
recursion function like usb_show_info() will get the
input parameter of USB device with NULL pointer, which
caused access to NULL pointer and cause CPU exception
and reset.
The patch fixed the issue via adding NULL pointer check.
JIRA number: A3700-714

Change-Id: I698c7585580731bffac4d6cdbfae93338b4e7b3c
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39075
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>