pinctl: a3700: set static mpp configuration from dts file
Due to the lack of pinctl driver, the MPP settings are hard
coded in the board specific code now but it is dedicated to
the devel-board. In order to faciliate the customer board
porting, this patch refined the existing WA by retrieving
register value of GPIO PINs Selection Registers from dts file
and updated dts files for espressobin board and devel board
accordingly.

Please note: The register value defined in dts is not bit(s)-
to-pin mapped configurations. For the details, please refer
to the following register definitions.
- north bridge: 0xD0013830
- south bridge: 0xD0018830

This patch will be replaced by the formal pinctl driver in
future.

Change-Id: Iad4867056b619fa62ffcdcd14b0e104127d5e61b
Signed-off-by: zachary <zhangzg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39133
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>