fix: bubt: mvebu: set correct usb devnum for image read
When reading image file from usb device, the devnum is set
to a incorrect value - UCLASS_USB. The first active usb
device is supposed to be used as 'bubt' source, thus the
devnum should be set to '0'.
Of all the available USB storage devices on multiple XCHI/
EHCI ports, the first device enumerated is selected.
In case there are two devices attached to EHCI and XHCI
port simultanously, the device on the XHCI port will be
selected.

Change-Id: I25ffadb0ac0b614f0a4bc192f123190064c188d5
Signed-off-by: allen yan <yanwei@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39262
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: Wilson Ding <dingwei@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>