phy: marvell: a3700: align comphy lane number in dts/driver with func spec
In current u-boot dts and comphy driver, lane0 is for PCIe/SGMII0 and
lane1 is for USB3/SGMII1 which opposites the A3700 spec.
This patch aligns lane number with the spec - lane0 is for USB3/SGMII1
and lane1 is for PCIe/SGMII0.
Please note that the generic API comphy_mux_init takes it for granted
that lanes' mux bits in selector base register are ordered by lane
number, but for a3700 lane1 mux bits is before lane0, this patch
re-orders serdes map and mux data and passes them to the API.

Change-Id: Iec744792a264f532cfb30b8a718286702502dbfa
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/54457
Reviewed-by: Igal Liberman <igall@marvell.com>
Tested-by: Igal Liberman <igall@marvell.com>