Clearfog Base: support m.2 PCIe cards
The latest board revision of the ClearFog Base updates the m.2
connector to also support certain types of PCIe cards.  Add the
proper ifdef to allow this to be configured at boot.  This commit
adds the ifdef in the same style as the other configuration changes
such that SATA is the default and it can be configured to be PCIE
by changing the ifdef 1 to ifdef 0.

Signed-off-by: Jon Nettleton <jon@solid-run.com>