switch: mv88e6xxx: add port setup in initialization
With default setting in port control register, switch works well; but
linux kernel will change the port control setting, for example, it will
add tag for egress frames in cpu port; then after reboot in kernel,
the changed port control setting is kept in u-boot, but the changed
setting does not work for switch in u-boot.
This patch adds port setup in initialization to make sure that switch
can work properly in u-boot.

Change-Id: I94acb01c97800189b0be1f5099460d73a8d0a567
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40752
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yelena Krivosheev <yelena@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>