efuse: a3700: Add DEVICE_DEPLOYMENT support
In order to put the chip into trusted mode, we need to set the
DEV_DEPLOY bits in efuse. Add this support for A3700.
The AES KEY in OTP will be masked after setting these bits.
The difference between trusted mode and non-trusted mode can be
found in the A3700 BootRom Reference Manual

Change-Id: Ie28112318df91a8feea5715ba25457034cf50e7a
Signed-off-by: zachary <zhangzg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40838
Reviewed-by: Terry Zhou <bjzhou@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>
Tested-by: Hua Jing <jinghua@marvell.com>