usb: hub: Add a new API to test if a hub device is root hub
Sometimes we need know if a given hub device is root hub or not.
Add a new API to test this. This removes the xHCI driver's own
version is_root_hub() and change to use the new API.

While we are here, remove the unused/commented out get_usb_device()
in the xHCI driver too.

Change-Id: Ifc4138e7b0cf62c57badbb20fe4c7ebf867f68f4
Signed-off-by: Bin Meng <bmeng.cn@gmail.com>