arm64: mvebu: config: move env to the end of the 4MB boot device
The size of flash_image.bin with the tee binary exceeds 1.5MB.
The tee binary size can grow up to 2MB, and without the environment
relocation leads to a boot image  corruption upon running saveenv.
Relocate the u-boot environment to the end of a minimum-sized
boot device (flash or eMMC boot partition of 4MB).

Change-Id: Ie829e0c305e18d2392c0b2bfca8f64cb9a122bb3
Signed-off-by: Zhoujie Wu <zjwu@marvell.com>
Signed-off-by: wangwen <wangwen@marvell.com>
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42647
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>