efuse: a38x: add DM and dts support for efuse driver
Efuse feature is fixed with hardcoded for a38x efuse, which is not
flexible for different mvebu SoC chips.
This patch adds driver model and dts for efuse driver and efuse
dts node in dtsi file and enables it on dts file.
Besides, it removes those older mvebu fuse codes.

Change-Id: I745ee5b3c4c5201671744d480585c59aaf6be81a
Signed-off-by: zachary zhang <zhangzg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/54568
Reviewed-by: Igal Liberman <igall@marvell.com>
Tested-by: Igal Liberman <igall@marvell.com>