fix: config: a3700: enable edit for ethaddr env
Without CONFIG_ENV_OVERWRITE, environment variable 'ethaddr'
could not be overwritten at all.
This patch enables it for Armada3700, and since
there has already been this define for Armada7/8K in
mvebu_armada-8k.h, so move it to mvebu_armada-common.h,
which is shared by Armada7/8K and Armada3700.

Change-Id: I4764267c322a26c1ccdf406433c9a3b857725dd5
Signed-off-by: jinghua <jinghua@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43452
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>