fix: net: overwrite ethaddr env with mac address in ROM
Now for eth devices that have their own ROM with valid Mac Address,
u-boot uses the Mac Address in ethaddr env, and Linux uses the
one in eth device's ROM. This could confuses some switch devices,
and breaks the networking access.
To fix the issue, this patch aligns u-boot and linux's behaviour
in this case, to use the Mac Address in eth device's ROM all the
time.

Change-Id: I89ed7f43bbe7c6e4451eeea85861c58514e820dd
Signed-off-by: jinghua <jinghua@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43753
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>