net: always use MAC address stored in the eth device's ROM
When u-boot environment ethaddr is changed by 'editenv' or 'setenv',
callback routine in ethernet uclass driver would update the ethaddr
in eth device's platdata, to keep it align with u-boot env all the
time. But for the eth device that has got its own rom with valid
mac address in it, it should stick to its own mac all the time,
this behavior is aligned with Kernel.
This patch updates the callback routine, for devices that have
registered 'read_rom_hwaddr' ops, which means have valid mac
address in its own ROM, do not set ethaddr in platdata according
to new u-boot env.

Change-Id: Idc2be624977a33130bdf827ab0ce7b04b939d606
Signed-off-by: jinghua <jinghua@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43754
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>